Do something with Bugzilla's automatic links
In Bugzilla comments the text 'bug NNNNNN' is automatically linked to
that bug. (We can't make it do the same with GitLab issues, but we can at
least link back to the Bugzilla bug.)

Bugzilla also links 'attachment NNNNNN' and 'comment #NN' which we don't
handle, but we remove the '#' so that GitLab doesn't make a spurious link
to an unrelated issue.

Closes #21, #22.