Analyze bugzilla comments to produce more friendly output
This tries to produce more readable GitLab comments by analyzing the text
of the Bugzilla comment. For example, if a comment starts with "Created
attachment NNNNNN" and the attachment is a patch, then the GitLab comment
will start with a tools emoji and the text "(user) submitted a patch:"

Another example is when a comment starts with "Review of attachment
NNNNNN" the GitLab comment will start with a magnifying glass and the
text "(user) reviewed patch NNNNNN:", and the diff blocks in the comment
will be highlighted as such.