Map keywords to standard labels where possible
Some of the newly created group labels map directly to existing
keywords, so simply translate them to the correct label.

There are more keywords that relate to labels, but the mapping is
a lot more fuzzy:

 - `usability` and `ui-review` can map to `2. Needs Design` in some
    cases, but not in others
 - `documentation` can be either `8. Developer Docs` or `8. User Docs`
 - all of `I18N`, `L10N` and `string` may map to `8. Translation`, but
    not always

Keep it safe and leave those out for now.

Fixes https://gitlab.gnome.org/External/bugzilla-to-gitlab-migrator/issues/4.