Add list of dependencies and blockers
While GitLab does not have a way to express dependencies between issues,
we still want to preserve that information coming from Bugzilla.

The simplest way to do so is to add the list of dependent and blocked
bugs to the issue description; to avoid excessive work, we don't inspect
the bugs for their own status, or recursively descend to create a tree
of dependencies.

Closes #8