Quote backtraces in comments
The code was adapted from
http://cpansearch.perl.org/src/MKANAT/Parse-StackTrace-0.08/lib/Parse/
which was used in the extension we had on Bugzilla that parsed
backtraces. This code can be a lot simpler since it only needs to
identify the beginning and end of the backtrace.

The backtrace quoting code is standalone so let's put it in a separate
file and test it for ease of development.

Closes #1