Quote XML-like tags when not already in backticks
This looks for XML-like tags such as <tag> or <some_tag attr="foo">
and quotes them in backticks if they are not already in backticks.

I'm not sure the regex is the most maintainable solution, but
iterating through the string and keeping track of the backticks state
is probably equally cumbersome.

Closes #17.