Cache users including secondary emails
Some people were using a different email for Bugzilla than for their
LDAP account, which is the account used in GNOME for regular
contributions permissions.

We are telling these people to add as their secondary email the email
they were using in Bugzilla so we can subscribe them to the issues they
were subscribed in Bugzilla.

Unfortunately, the API for python-gitlab and GitLab itself can query
users based in their primary email, but not based on their secondary
emails.

This commit adds a first phase of downloading the users data into a file
so we can use it as a users cache, including their secondary emails.