Fix setup.py to install the whole package
It should have used the `packages` argument instead of `py_modules`. This
worked in my local setup for whatever reason, but fails when you actually
try to install the package on a clean system.

We add a sanity check stage to CI to test exactly this: installing the
package on a clean system and running `bztogl --help`.