Allow uncached comment authors
We do not cache Bugzilla users that match the default assignee or the
default QA contact for a component or a product, because the GNOME
Bugzilla instance typically uses fake addresses for those fields. Sadly,
some products over the years decided to use real people as default
assignees and QA contacts, and those people may not have used an email
address that is mapped on any of the GNOME infrastructure services.

This means that as soon as we get a comment from a person who's the
default assignee for a specific bug, we get an empty entry in the users
cache, and it breaks the migration script.